move deleteshipment test to homepage, remove unused function
